Store `extras` in struct
This alternative implementation makes it easier to construct arrays
where indexing depends on some additional data without needing to
fear non-constant globals.  This comes at the cost of needing to
provide an extra argument to the supplied `getindex` function.

It also passes the index as a tuple. This makes it more likely to be
automatically specialized than the `Vararg` would be.